Unordered maps (#47)
* refactor, indices folder and change maps to unordered
