+++ .github/workflows/cppcheck.yml
@@ -11,7 +11,7 @@ jobs:
     - uses: actions/checkout@v1
     - uses: actions/setup-python@v1
     - run: pip install cpplint
-    - run: cpplint --recursive --exclude id_index/bloom_filter.hpp .
+    - run: cpplint --recursive --exclude indices/make_sequence_bloom_tree/bloom_filter.hpp .
   build:
     runs-on: ubuntu-latest
     steps:
