+++ graph_query.h
@@ -28,8 +28,8 @@
 #include <boost/graph/vf2_sub_graph_iso.hpp>
 #include <boost/date_time/posix_time/posix_time.hpp>
 #include "get_traces_by_structure.h"
-#include "folders_index_query.h"
-#include "query_bloom_index.h"
+#include "indices/folders_index_query.h"
+#include "indices/query_bloom_index.h"
 #include "common.h"
 
 
@@ -61,7 +61,7 @@ std::string get_service_name_for_node_index(
 );
 fetched_data fetch_data(
     traces_by_structure& structs_result,
-    std::map<std::string, std::vector<std::string>>& object_name_to_trace_ids_of_interest,
+    objname_to_matching_trace_ids& object_name_to_trace_ids_of_interest,
     std::vector<query_condition> &conditions,
     gcs::Client* client
 );
