+++ graph_query_main.cc
@@ -1,21 +1,22 @@
 #include "get_traces_by_structure.h"
 #include "graph_query.h"
-#include "query_bloom_index.h"
+#include "indices/query_bloom_index.h"
 
 int main(int argc, char* argv[]) {
     // query trace structure
     trace_structure query_trace;
-    query_trace.num_nodes = 1;
+    query_trace.num_nodes = 3;
     query_trace.node_names.insert(std::make_pair(0, "frontend"));
-    // query_trace.node_names.insert(std::make_pair(1, "adservice"));
-    // query_trace.node_names.insert(std::make_pair(2, ASTERISK_SERVICE));
+    query_trace.node_names.insert(std::make_pair(1, "adservice"));
+    query_trace.node_names.insert(std::make_pair(2, ASTERISK_SERVICE));
 
-    // query_trace.edges.insert(std::make_pair(0, 1));
-    // query_trace.edges.insert(std::make_pair(1, 2));
+    query_trace.edges.insert(std::make_pair(0, 1));
+    query_trace.edges.insert(std::make_pair(1, 2));
 
     // query conditions
     std::vector<query_condition> conditions;
 
+    /*
     query_condition condition1;
     condition1.node_index = 0;
     condition1.type = bytes_value;
@@ -27,6 +28,7 @@ int main(int argc, char* argv[]) {
     condition1.property_name = "trace-id";
 
     conditions.push_back(condition1);
+    */
 
     // querying
     auto client = gcs::Client();
@@ -40,7 +42,7 @@ int main(int argc, char* argv[]) {
     ret.func = ret_union;
     boost::posix_time::ptime start, stop;
 	  start = boost::posix_time::microsec_clock::local_time();
-    auto res = query(query_trace, 1661969609, 1661969612, conditions, ret, false, &client);
+    auto res = query(query_trace, 1663255784, 1663255800, conditions, ret, false, &client);
     // auto res2 = get_traces_by_structure(query_trace, 1660239561, 1660239571, &client);
     stop = boost::posix_time::microsec_clock::local_time();
     boost::posix_time::time_duration dur = stop - start;
