+++ indices/folders_index_query.cc
@@ -1,6 +1,6 @@
 #include "folders_index_query.h"
 
-std::map<std::string, std::vector<std::string>> get_obj_name_to_trace_ids_map_from_folders_index(
+objname_to_matching_trace_ids get_obj_name_to_trace_ids_map_from_folders_index(
 	std::string attr_key, std::string attr_val, int start_time, int end_time, gcs::Client* client
 ) {
 	std::vector<std::future<std::unordered_map<std::string, std::vector<std::string>>>> response_futures;
@@ -24,7 +24,7 @@ std::map<std::string, std::vector<std::string>> get_obj_name_to_trace_ids_map_fr
 			object_metadata->name(), bucket_name, start_time, end_time, client));
 	}
 
-	std::map<std::string, std::vector<std::string>> response;
+	objname_to_matching_trace_ids response;
 	for_each(response_futures.begin(), response_futures.end(),
 		[&response](std::future<std::unordered_map<std::string, std::vector<std::string>>>& fut) {
 			std::unordered_map<std::string, std::vector<std::string>> obj_name_to_trace_ids_map = fut.get();
