+++ indices/make_folders_index/trace_attributes.cc
